Remove autobuilt reference docs and update doc config
+ Removed all of the automatically creatable reference docs for both the BuiltIn commands for
the bot as well as the actual exposed functions. These should be generated automatically by
PlatyPS during the build and don't need to be in place for the actual repository.
+ Updated the MkDocs config to include site discription and author for HTML metadata, added
the edit url for an 'edit on github' experience (this won't work on the reference docs, btw,
if we don't keep them source controlled and just build them as needed), and versioned the
documentation as