Pass the backend type into a Command
Some systems like Teams will not automatically expand some output, like image links
so sending them markdown instead of a plain link will correct this
to enable this behaviour we need to be able to let other plugins know which
backend it is talking to.
Using $Global:PoshbotContext.BackendType will enable this.