(GH109) Use CommandParser to strip out extra line breaks when testing if a message is a bot command
When text is pasted into Teams, it will prepend \r\n to the message that we receive. This trips PoshBot up when it evaluates if the message is a bot command or not. This commit runs the message through the CommandParser class to strip any lines breaks out so we're left with what the user actually entered.