(GH132) Zero out channel rules when adding new ones
By default, the ChannelRules property of the [BotConfiguration] object has a single channel rule that allows all commands in all channels. This is so that when creating a new configuration object with New-PoshBotConfiguration and specifying no addition parameters, we have a working configuration.

When creating a new bot configuration object AND passing it one or more channels rules, we should zero out the array as it is implied by passing in channel rules that the rules should be exactly what was passed in and not IN ADDITION to the default one.